refactor: split CrateStrategies into one file per strategy class
Move Picks, NewArrivals, Thematic, and HiddenGems each into their own
file under app/services/crate_strategies/. Requires moved to Thematic
file where Digest::MD5 is used. Autoloading confirmed via eager_load!.
